+++ CHANGELOG.md
@@ -15,6 +15,9 @@ All notable changes to MiniJinja are documented here.
 - Added `mj_value_new_bytes` to the C-ABI.  #749
 - Added `mj_value_as_bytes` to the C-ABI to borrow from strings or
   byte values.  #750
+- Fixed buggy `mj_err_get_detail` and `mj_err_get_template_name`.  These
+  did not work correctly.  To fix them the return value now needs to be
+  freed.  #754
 - Fix a compilation issue on 32bit systems when `AtomicU64` is
   not available in minijinja-contrib.  #755